change destructible obj implementation from scriptable obj to prefabs
