+++ CrimeWave/Assets/Resources/Prefabs/DestructibleObjects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1afebcd3f1937a849b8aa08b73569db5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
