+++ CrimeWave/Assets/Resources/Prefabs/DestructibleObjects/BigCrate.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8bd5358a1eee18b44bbd10082d7403d1
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
