+++ CrimeWave/Assets/Resources/Prefabs/DestructibleObjects/SmallCrate.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a6cb250e483eff745bd0c226f6698891
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
