--- CrimeWave/Assets/Resources/ScriptableObjects/DestructibleObjects.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f8364708803f8304f9fff014efa572d0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
