--- CrimeWave/Assets/Resources/ScriptableObjects/DestructibleObjects/BigCrate.asset
@@ -1,18 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ac7db6834ae5e0f45891b3832f8b3733, type: 3}
-  m_Name: BigCrate
-  m_EditorClassIdentifier: 
-  objectName: Big Crate
-  sprite: {fileID: 1407484440448051338, guid: f0341a37f08163447978f0ef0be0aa78, type: 3}
-  maxHealth: 50
-  lootDropAmount: 500
