--- CrimeWave/Assets/Resources/ScriptableObjects/DestructibleObjects/BigCrate.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f27f64f75a63e9a4ba2076a703524042
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
