--- CrimeWave/Assets/Resources/ScriptableObjects/DestructibleObjects/SmallCrate.asset
@@ -1,18 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ac7db6834ae5e0f45891b3832f8b3733, type: 3}
-  m_Name: SmallCrate
-  m_EditorClassIdentifier: 
-  objectName: Small Crate
-  sprite: {fileID: -8605155236358719273, guid: 9cb8e447e5a1b0c47bc75c7d22235f34, type: 3}
-  maxHealth: 20
-  lootDropAmount: 100
