--- CrimeWave/Assets/Resources/ScriptableObjects/DestructibleObjects/SmallCrate.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 42ea489e1b9f2404e90779673bc9faa0
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
