--- CrimeWave/Assets/Resources/Sprites/Destructible Objects.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: fe4dbafd432cf3a45b8e08e1a6647a17
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
