+++ CrimeWave/Assets/Resources/Sprites/DestructibleObjects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fe4dbafd432cf3a45b8e08e1a6647a17
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
