+++ CrimeWave/Assets/Scripts/DestructibleObject.cs.meta
@@ -1,2 +1,2 @@
 fileFormatVersion: 2
-guid: ac7db6834ae5e0f45891b3832f8b3733
+guid: 22a2ebdb1c063e94b8b731bd5ab075e8
