--- CrimeWave/Assets/Scripts/DestructibleObjectBehavior.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 22a2ebdb1c063e94b8b731bd5ab075e8
